+++ packages/ui/src/lib/i18n/index.ts
@@ -1,5 +1,17 @@
 export * from './i18n.types';
 export { EAGAMI_I18N_CONFIG, provideEagamiUi } from './i18n.provider';
 export { EagamiI18nService } from './i18n.service';
-export { EAGAMI_MESSAGES, el, en, esES, frFR, pl } from './messages';
+export {
+  EAGAMI_ALL_LOCALES,
+  de,
+  el,
+  en,
+  esES,
+  frFR,
+  is,
+  nl,
+  pl,
+  ptBR,
+  zhCN,
+} from './messages';
 export { frenchSpacing } from './french-spacing';